Rename useFetchData to useFetchPagedData
